agent/requirements.txt
pytest
paramiko
requests
allure-pytest
